restructuring w/ other files
